New package: `vdhcoapp`
Package the native companion app for the Video DownloadHelper browser
addon (https://www.downloadhelper.net/).  This companion app is required
if the site uses the HLS or DASH streaming protocol, or if `ffmpeg` is
needed to process the video (either to combine separate video and audio
streams, or to perform some actual conversion with reencoding).

The companion app is written in JavaScript for Node.js and was intended
to be packaged into a single executable file using the `pkg` Node.js
package, but that method is not really compatible with NixOS (the stock
`node` executable does not work like that without some patches, and the
prebuilt Node.js interpreter executables provided by `pkg` do not work
in NixOS), therefore some patches were needed to make the code run from
raw script files without packaging.