Make the repository also usable as a flake
The flake exposes individual packages under `packages.<system>` (using
the `nixpkgs` input as the base).  In addition, the overlay which adds
all those packages is exposed as `overlay` (it is the same overlay as
returned by `overlay.nix`, so it does not put the packages into any kind
of namespace and can potentially overwrite packages which came from
other sources).