Update the build workflow for Nix 2.4
The default Nix version installed by `cachix/install-nix-action` was
updated to 2.4, and the `nix run` command has been significantly changed
in that version.  Update the command used to run `nix-build-uncached` to
make it compatible with Nix 2.4; this also requires adding the
`experimental-features = nix-command flakes` option to the Nix
configuration (the `nix-command` feature is also needed for
`nix-build-uncached` itself, because it uses `nix build`).

Upstream issue: nix-community/nur-packages-template#52

The comment about `nix-build-uncached 1.0.0` and the `nixos-unstable`
reference were no longer relevant (both nixos-21.05 and nixos-unstable
have the 1.1.0 version at the moment) and were removed.