ci: Trigger NUR update from a separate job
Triggering NUR update from the `tests` job results in multiple requests
for a single CI run (because that job uses the `matrix` feature to build
packages in multiple configurations), and sometimes this results in
multiple copies of the update workflow in the NUR repo running in
parallel (because the code that attempts to prevent that apparently has
a race condition).  Move the triggering to a separate job that gets run
only once to avoid the problem.