virt-manager-2.2.1: Fix build by backporting NixOS/nixpkgs#169221
The `setuptools` Python package in Nixpkgs was updated to version 61,
which contains a breaking change (now `packages=[]` is required in
`setup.py` for empty packages).  The `virt-manager` package in Nixpkgs
was updated in NixOS/nixpkgs#169221 to fix the build; backport the same
change to the copy of virt-manager-2.2.1.

The backport uses `postPatch` instead of `prePatch`, as suggested in the
review comment to the above PR which was added the PR was merged.