ci: Build the `packages` and `devShells` outputs of the flake
In addition to testing builds with various `NIX_PATH` values, also test
builds of flake outputs (which use locked revisions for all inputs), and
test `devShells` too.

The flake build part is factored out to a reusable workflow so that it
would be possible to add `x86_64-darwin` tests as a separate job.