Update support code for upcoming CI job changes
Replace both the flake code and the NUR entry points (`default.nix` and
`overlay.nix`) with the new implementation that supports a more advanced
mechanism for the CI job matrix generation.  Unlike the old `ci.nix`,
which supported only NUR packages, the new mechanism also supports flake
outputs like `packages`, `checks` and even `nixosConfigurations`, with
the possibility to split builds into multiple jobs running in parallel
if desired.

The new CI job code is not yet present (this is intentional, so that the
old `ci.nix` could be tested with the old CI job - it should still work
the same way after these changes).