lib: Add `concatMapAttrsToListRecursive`
The existing `flattenAttrs` function turned out to be not generic enough
for some use cases; in particular, getting the path to the attribute for
further processing would be possible only through converting the path to
a string and back again.  Extract the inner logic of that function into
a separate function (`concatMapAttrsToListRecursive`), then reimplement
`flattenAttrs` in terms of that function.