Add language pack packages for Firefox and Thunderbird
The `firefox`, `firefox-esr` and `thunderbird` packages in Nixpkgs are
properly built from sources (unlike `firefox-bin` and `thunderbird-bin`,
which are repackaged from binary tarballs built by Mozilla).  However,
while the binary packages also provide localized versions for various
languages, the packages built from sources do not support localization
out of the box and require going to addons.mozilla.org to install the
required language packs.

Add packages which contain various language packs for Firefox and
Thunderbird in the same format as other extension packages, so that
those language packs can be installed in the user profile as regular
extensions.  (This method of installation still requires the user to
confirm enabling the extension at the first startup, but that just
requires some clicks without needing to search for the particular
language pack, so it is still useful.)