ci: Use `cachix/install-nix-action` from the upstream repo
The `sigprof/install-nix-action` fork was used before to get better log
messages (https://www.github.com/cachix/install-nix-action/pull/139).
That PR has been merged now, so it is possible to switch to the upstream
repository for that action (but a release including that PR has not been
made yet, so a pin to the commit hash is used).