ci: Sanitize build log names for uploading
Nix build log names can sometimes contain characters like `?`, which are
not allowed in GitHub artifacts:

    Invalid characters include:  Double quote ", Colon :, Less than <,
    Greater than >, Vertical bar |, Asterisk *, Question mark ?,
    Carriage return \r, Line feed \n

Replace the mentioned characters (except the control characters `\r` and
`\n`, which probably should not be there anyway) with `_` (this should
not introduce collisions, because the derivation hashes should be unique
anyway).  Also replace the backslash character with `_`, because most
likely it would be interpreted as a path separator when uploaded into
the artifact.