Temporarily replace missing source for `tor-browser-bundle-bin`
Currently the flake is locked on an old nixpkgs snapshot on purpose, and
therefore the `tor-browser-bundle-bin` package refers to an old version
of Tor Browser; however, all main download locations already have that
version removed, therefore the package cannot be built.  Temporarily
replace the source with another URL where it is still present, so that
the package could be built and cached (then the modification could be
reverted again, but the package would remain usable as long as it
remains in the binary cache).