ci(auto-rebase): New workflow for rebasing PRs automatically
The PRs which have the `autorebase:opt-in` label will now be rebased
every time the `master` branch gets new commits, or every time the `CI`
workflow runs for any other reason.  A more traditional way is to run
the `auto-rebase` workflow on the `push` event, but that would result in
having the `auto-rebase` workflow in the list of commit checks, which is
annoying, and GitHub still does not offer any way to hide some workflows
from the list of checks, therefore this feature is not enabled here.

However, pushing a commit with the `skip-ci` flag to the `master` branch
will also skip the `auto-rebase` execution, therefore it may be
necessary to run this workflow manually in some cases.


skip-checks: true