ci(auto-update-flake): New workflow to update flake inputs
Add a new workflow which checks all flake inputs for possible updates
and creates PRs for any inputs that have updates.

For the `nixpkgs` and `nixos-unstable` inputs this workflow enables
automatic merge for the resulting PRs, so that updates would be applied
automatically if the CI tests succeed after those updates.  For other
flake inputs the PR must be reviewed and merged manually.

All generated PRs are also tagged with the `autorebase:opt-in` label, so
that they will be automatically rebased when the `master` branch gets
new commits (this includes the case when some of the generated PRs get
merged automatically).

Branch names for the generated PRs (`auto-update/flake/INPUT`) are also
handled specially by the `CI` workflow: all NUR checks are skipped, and
only flake tests are run, so that any NUR-specific failures don't block
updates for flake inputs.


skip-checks: true