ci(auto-rebase): Avoid concurrency for this workflow
Running multiple instances of the `auto-rebase` workflow in parallel may
be dangerous, because those instances may try to manipulate the same PR
at the same time.  Set `concurrency.group` to `${{ github.workflow}}` to
disable concurrency for this workflow.

Also set `concurrency.cancel-in-progress` to `false` to make sure that
any running instances of the `auto-rebase` workflow never get cancelled
(it is better to wait until the previous workflow finishes than risk
leaving some PR in some intermediate state).  In this mode only those
workflow instances that were in the `pending` state (waiting for another
instance to complete) can get cancelled; this should be safe, because a
`pending` workflow instance did not start any operations yet.


skip-checks: true