mozilla-langpack: Build a fixed output derivation if possible
If `sources.json` contains a precomputed NAR hash for the matching
install path, build the language pack as a fixed output derivation
(using `fetchurl` with a custom `postFetch` script that moves the
downloaded file into the required location).  If `sources.json` does not
have the required NAR hash, fall back to the existing implementation
which would produce a regular derivation with the output path depending
on `pkgs.stdenvNoCC`.