mozilla-langpack: Add NAR hash generation to the update script
In addition to storing just the original file hashes, make the
`update.sh` script actually download the language pack files, check
their hashes, then install the language pack into the intended
subdirectory with the proper file name and calculate the NAR hash for
the resulting directory, and finally store both the installation path
and the NAR hash in the generated `sources.json`.  If the installation
path used by the Nix code would match the path used by the update
script, the language pack could become a fixed output derivation, saving
some space in the binary cache (normally derivations are different for
every supported Nixpkgs branch, but with a fixed output derivation the
only time when a new derivation would be needed is when the language
pack actually gets updated).