ci: Make all CI jobs respect cancellations
Using `if: always() && ...` without checking for `!cancelled()` results
in the job continuing to run even after the workflow was cancelled.  Add
checks like `(always() && !cancelled())` to all such jobs, so that any
cancel requests would be handled properly, according to the docs at:

    https://docs.github.com/en/actions/managing-workflow-runs/canceling-a-workflow#steps-github-takes-to-cancel-a-workflow-run

While at it, use `if: >-` to strip any extra newlines.