mozilla-langpack: Reuse NAR hashes for unchanged files
If a language pack file has the same hash as listed in the previous
version of `source.json`, and the installation path is also the same,
copy the NAR hash from the previous version of `source.json` instead of
downloading the file and recalculating the hash.  This makes the update
process much faster in the usual case when no updates are found.