virt-manager: Fix the availability-related meta attributes
Recently the `virt-manager` package was updated to make it build on
Darwin (https://www.github.com/NixOS/nixpkgs/pull/185415); this broke
the customized version of `virt-manager`, because it inherited the
updated meta attributes, but the old 2.x version of `virt-manager` still
does not build on Darwin.  Fix the problem by merging the values of
`meta.broken`, `meta.platforms` and `meta.badPlatforms` in the
appropriate way, so that the combined package gets built only if both
the original package and the old version could be built.