Add `lib.forceCached` to work around `allowSubstitutes = false`
Some derivations in Nixpkgs annoyingly have `allowSubstitutes = false`;
this option breaks `nix-build-uncached` (the derivation always gets
rebuilt even when the previous build had pushed its outputs to Cachix),
and is impossible to override in a clean way.  Copy the `forceCached`
workaround from `nix-build-uncached`; this function can be used to wrap
the problematic derivations into cached ones.

The implementation of `forceCached` uses `pkgs.coreutils`, therefore it
cannot be exported from the `lib` output of the flake; instead, it is
available under `legacyPackages.${system}.lib`.