hosts(test1): Use `kicad-small` from `nixpkgs`
The KiCad package fron `nixos-unstable` does not work properly in a
system based on the stable `nixos-22.05` branch for some reason,
therefore the actual host configuration has been changed to use
`pkgs.kicad-small` instead of `nixos-unstable.kicad-small` some time
ago.  Update the test host configuration to match that change.

This also makes it possible to update the `nixos-unstable` branch to a
recent version (apparently the KiCad package does not build there at the
moment, and the test catches that breakage and prevents the update).