flake: Make the `pre-commit-hooks.nixpkgs-stable` input follow `nixpkgs`
The `pre-commit-hooks` flake now has two Nixpkgs inputs (`nixpkgs`,
which originally points to a snapshot of the `nixos-unstable` branch,
and `nixpkgs-stable`, which currently points to a snapshot of the
`nixos-22.05` branch).  Make the `nixpkgs-stable` input follow our
`nixpkgs` input instead of using an obsolete snapshot, so that the
number of different Nixpkgs versions in use does not grow.

(It could be possible to make the `pre-commit-hooks.nixpkgs` input to
follow `nixos-unstable` instead, but currently that is not needed,
because `pre-commit-hooks` was fixed to not fail on `nixos-22.05` unless
the packages which are missing in that release are actually used by some
checks, and the `pre-commit-hooks.nixpkgs` input is also used to provide
`pre-commit-hooks.gitignore.nixpkgs`.)