ci: Switch to the latest snapshot of `cpcloud/flake-update-action`
The upstream version of `cpcloud/flake-update-action` now includes the
required updates to avoid deprecation warnings.  Unfortunately, the
`v1.0.4` release of that action does not contain all required updates
(in particular, it still refers to `peter-evans/create-pull-request@v3`
and `peter-evans/enable-pull-request-automerge@v1`, which use the
obsolete `node12` version instead of `node16`), therefore the latest
commit from the `main` branch is used instead of the release tag.