Replace `hardware.video.hidpi.enable` with `sigprof.hardware.video.uiScale`
The `hardware.video.hidpi.enable` option was removed in NixOS 23.05:

  https://www.github.com/NixOS/nixpkgs/pull/222236
  https://www.github.com/NixOS/nixpkgs/pull/222689

However, some kind of UI scaling option is needed to select the
appropriate console and GRUB fonts (even if the kernel has some kind of
automatic detection for HiDPI now, the builtin font does not support
Cyrillic characters).  Because there does not seem to be a good option
for that in Nixpkgs (`services.xserver.dpi` is not really suitable), a
custom `sigprof.hardware.video.uiScale` option is introduced instead.
This option specifies the UI scale factor instead of the DPI value,
because the desired UI pixel size for the same physical DPI value may
actually vary depending on the device form factor.