Fix `sigprof.hardware.gpu.driver.nvidia.legacy_340` on NixOS 23.05
Before NixOS 23.05 the `legacy_340` NVIDIA driver could not be built for
Linux kernels newer than 5.4.x; however, in NixOS 23.05 that driver was
patched to support kernels up to 6.1.x (using patches from AUR):

  https://www.github.com/NixOS/nixpkgs/pull/211524

Unfortunately, those patches apparently had broken the driver build for
Linux 5.4.x, therefore the code that downgraded the kernel to 5.4.x when
the `legacy_340` driver was enabled now needs to be turned off if the
NixOS release is 23.05 or newer.

The change was not actually tested on the real hardware, because the old
machine which needs the `legacy_340` driver is no longer in use.