ci: Bump Nix to 2.18.1
Change `CI_NIX_INSTALL_URL` to point to Nix 2.18.1.  This should not
break anything, because the old `cpcloud/flake-update-action` should
still be compatible with that version, and the `nix-build-uncached`
package in the locked `nixpkgs` input (NixOS 23.11) is updated for
compatibility with Nix >= 2.15.

Note that the input of `nix-build-uncached` does not need to have the
`^*` suffixes added to `*.drv` paths, because that input is passed to
`nix-build`, which did not have incompatible changes in Nix 2.15 (the
changes were made to `nix build`, which is used internally by
`nix-build-uncached` after getting the list of derivations that needs to
be built from `nix-build --dry-run`).