virt-manager-2.2.1: Limit Python version to < 3.12
Recently `nixpkgs-unstable` bumped the default `python3` version to
3.12.x, but the old virt-manager 2.2.1 code is not compatible with
Python >= 3.12, because it uses the obsolete `imp` module which got
removed in Python 3.12.  Check the Python version used by `nixpkgs` and
switch to Python 3.11 if the default version is 3.12 or newer (but keep
using the default Python version if it's older than 3.11 to avoid
potential breakage on older `nixpkgs` branches).