mozilla-langpack: Derive langpack names from the apparent app name
The code was trying to derive the langpack package names from the
attributes of the app package; in particular, it added the `-esr`
component if the version of the app package had the `esr` suffix.
That broke when NixOS 24.05 switched the `thunderbird` package to
point to the 128.x ESR track, which has the `esr` suffix in its
version number, unlike the previous 115.x track.

Instead of trying to be clever with the ESR status, just derive the
langpack package names from the name of the main package as it appears
in Nixpkgs, so that `thunderbird` gets `thunderbird-langpack-XX` and not
`thunderbird-esr-langpack-XX`, even when the version of the
`thunderbird` package has the `esr` tag.