Drop the `sigprof.hardware.gpu.driver.nvidia.legacy_340` option
NixOS/nixpkgs#320228 removed the `hardware.opengl.setLdLibraryPath`
option, which was required for the `legacy_340` NVIDIA driver to work
properly (NixOS/nixpkgs#120602).  The machine which needed the
`legacy_340` driver is no longer in use for quite some time, so there is
no way to test any possible fixes; just drop the obsolete option.