ci: Tune `wimpysworld/nothing-but-nix` parameters for speed
The default `hatchet-protocol: cleave` is too slow, because it removes
`/usr/local` synchronously, and that takes about 2 minutes on current
GitHub runners.  Switch to `hatchet-protocol: carve`, which just
combines the available space from the `/mnt` and `/` filesystems, and
does not remove any files from the runner; the resulting amount of disk
space should still be enough for the expected usage.